Fixes and improvements
We now use a yyyy/mm/dd directory structure for the photos output.
This should be easy to use (and scalable) on cloud storage.

There's a flag for committing changes, by default the program runs in
"dry run" mode. This allows the user to spot problems before actually
rearranging files on their computer.

Fixes and tidying.